requirements.txt
fastapi
SQLAlchemy
python-dotenv
uvicorn
pydantic
psycopg2
pypdf
groq
